Extracting WpComApiClient construction to WpComApiClientProvider to make the repo testable
