+++ WordPress/src/main/java/org/wordpress/android/support/aibot/repository/AIBotSupportRepository.kt
@@ -2,48 +2,31 @@ package org.wordpress.android.support.aibot.repository
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import okhttp3.OkHttpClient
 import org.wordpress.android.fluxc.utils.AppLogWrapper
+import org.wordpress.android.networking.restapi.WpComApiClientProvider
 import org.wordpress.android.support.aibot.model.BotConversation
 import org.wordpress.android.support.aibot.model.BotMessage
 import org.wordpress.android.util.AppLog
 import rs.wordpress.api.kotlin.WpComApiClient
-import rs.wordpress.api.kotlin.WpHttpClient
-import rs.wordpress.api.kotlin.WpRequestExecutor
 import rs.wordpress.api.kotlin.WpRequestResult
 import uniffi.wp_api.AddMessageToBotConversationParams
 import uniffi.wp_api.BotConversationSummary
 import uniffi.wp_api.CreateBotConversationParams
 import uniffi.wp_api.GetBotConversationParams
-import uniffi.wp_api.WpAuthentication
-import uniffi.wp_api.WpAuthenticationProvider
-import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 private const val BOT_ID = "jetpack-chat-mobile"
-private const val READ_WRITE_TIMEOUT = 60L
-private const val CONNECT_TIMEOUT = 30L
 
 class AIBotSupportRepository @Inject constructor(
     private val appLogWrapper: AppLogWrapper,
+    private val wpComApiClientProvider: WpComApiClientProvider,
 ) {
     private var accessToken: String? = null
     private var userId: Long = 0
 
     private val wpComApiClient: WpComApiClient by lazy {
         check(accessToken != null || userId != 0L) { "Repository not initialized" }
-
-        val okHttpClient = OkHttpClient.Builder()
-            .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
-            .readTimeout(READ_WRITE_TIMEOUT, TimeUnit.SECONDS)
-            .writeTimeout(READ_WRITE_TIMEOUT, TimeUnit.SECONDS)
-            .build()
-
-        WpComApiClient(
-            requestExecutor = WpRequestExecutor(httpClient = WpHttpClient.CustomOkHttpClient(okHttpClient)),
-            authProvider = WpAuthenticationProvider.staticWithAuth(WpAuthentication.Bearer(token = accessToken!!)
-            )
-        )
+        wpComApiClientProvider.getWpComApiClient(accessToken!!)
     }
 
     fun init(accessToken: String, userId: Long) {
